LayoutTests/platform/win/fast/xmlhttprequest/xmlhttprequest-multiple-sync-xhr-during-unload-expected.txt

frame "<!--frame1-->" - has 1 onunload handler(s)
CONSOLE MESSAGE: line 9: XMLHttpRequest cannot load xmlhttprequest-responsetype-json.json
CONSOLE MESSAGE: line 9: XMLHttpRequest cannot load xmlhttprequest-responsetype-json.json
CONSOLE MESSAGE: line 9: XMLHttpRequest cannot load xmlhttprequest-responsetype-json.json
CONSOLE MESSAGE: line 9: XMLHttpRequest cannot load xmlhttprequest-responsetype-json.json
CONSOLE MESSAGE: line 9: XMLHttpRequest cannot load xmlhttprequest-responsetype-json.json
Makes sure that we throw when doing sync XHRs during unload.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS xhrExceptionCount is 5
PASS successfullyParsed is true

TEST COMPLETE
